requirements.txt
antlr4-python3-runtime==4.9.3
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
fonttools==4.60.0
fsspec==2025.9.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
omegaconf==2.3.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pyarrow==21.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
tzdata==2025.2
wheel==0.45.1
